requirements.txt
python-gitlab
gitpython
